Fix floating ip timeout issues with uwsgi enabled
This reverts commit d4c9cc89f199899b20cfaea975ecd15ee66d3523 which
introduced a waiter to fix a bug in Tempest. Unfortunately that broke
neutron-tempest-with-uwsgi job where some of the tests started to fail
due to timeouts - waiting for the floating ip to get attached to a server
prolonged the tests which caused the timeout (it seems the ip attachment
process takes longer on an environment with uwsgi).
The failing tests also didn't require an attached floating ip or they
did however long enough after the server creation thus it is redundant
to wait for the ip attachment at the beginning (when the resource is
created) when it can happen "in parallel".

This patch moves the waiter from the common file to the test class
where the waiter is needed to fix a bug [1].

[1] https://bugs.launchpad.net/tempest/+bug/1945495

Closes-Bug: #1960022
Change-Id: Iddfa4ef874328fdee5418de5a8b1112007541883