Use stable constraint in tox to release new tag for Yoga
We are going to release Tempest new tag to declare the start of
support for stable/yoga. So that new tag use stable/yoga
constraint in the tox env.

For example: if anyone use Tempest new tag (30.0.0) in future say
1 year later then tox env also should use the yoga contstraint for
compatibility instead master.

These need to be move back to master constraint once release is done.

By using master constraint in tox for all the old tag is problem and
we faced the issue of constraint incompatibility due to that and we
have to work around it all over the devstack, tempest role, grenade to
set the stable constraint cia env var.

Change-Id: Ie3159da1b9b73652f52cf22d87ab1dd0ae6411d9