Fix compute_unified decoration
Current test class decorate the class with skip.
The problem is that when feature is disabled the skip section
returns a function type.
The ServersQuotaTest class inherits from a function type instead
type (class) and it fails with TypeError: function() argument

We dont see the exception because on false we dont try to run

testtool version is 2.5.0

Change-Id: I3ff2e59ca1eaae4f9b1eb1fb7148cd3efcbcd8f2