Make rescue, volume attachment compute tests to use SSH-able server
As you can see the details in bug#1960346, volume detach fails
on centos 9 stream while server is not fully booted. This
commit makes sure that erver creation as well as after the unrescue
server test wait for the server to be SSH-able before test start
performing the detach operation in cleanup.

Related-Bug: #1960346
Change-Id: Ib21a764e3cf81d761d738bf7bfaf5c264b739a16