Use UPPER_CONSTRAINTS_FILE for stable/ussuri testing
Tempest is going to be pinned for stable/ussuri(depends on) which
need to modify run-tempest role to use the constraints set by
the devstack or any job.

Because of Tempest 26.1.0 and with stestr 2.5.1 or 3.0.1 (beacause of upper
constraints of stestr 2.5.1 in stable/train and 3.0.1 in stable/ussuri )
which does not have new args exclude-regex so let's fallback to old arg if
new arg is passed.

Change-Id: I89ae6bd5137e57cb6ec5bc3c9edecc29489bca94