Add verification when restoring backups
If you want to restore a backup, specifying only the backup ID, you will
get a new volume. However, if you pass the volume_id argument, you are
targeting the restore of an existing volume.

The problem I am having is that restore_backup() adds a second call to
addCleanup to an existing volume and this results in a 400 HTTP error.

To fix this problem, I add a simple check when we want to use an
existing volume without calling cleanup.

Change-Id: I521f69091cdad60edfe4ed85368df5decc0a7f55