Use UPPER_CONSTRAINTS_FILE for stable/victoria testing
stable/victoria is in EM state and Tempest is going to be
pinned for stable/victoria which need to modify run-tempest
role to use the constraints set by the devstack or any job.

Because of Tempest 26.1.0 and with stestr 3.0.1 (beacause of upper
constraints of stestr 3.0.1 in stable/ussuri ) which does not have
new args exclude-regex so let's fallback to old arg if new arg is
passed.

Change-Id: I9cad0ccb02ee7bef0f94a1ee40e0fa117d89b5eb