Drop py3.6 and py3.7 from Tempest
In zed cycle, OpenStack projects and oslo lib
has dropped the py3.6 and py3.7 support

- https://governance.openstack.org/tc/reference/runtimes/zed.html

With oslo not supporting the py3.6 and py3.7, Tempest master is
failing and cannot be run on py3.6|7 env.

- https://bugs.launchpad.net/tempest/+bug/1975036

We discussed about how to solve it in ML[1] and QA meeting[2] and
decided to drop it from Tempest too. Stable branches testing from
stable/victoria to stable/yoga (currently supported branch by Tempest
master) can be tested on py3.8 or if cloud run those releases on py3.6
then Tempest can be run in virtual env with py38.

Closes-Bug: #1975036

[1] http://lists.openstack.org/pipermail/openstack-discuss/2022-May/028622.html
[2] https://meetings.opendev.org/irclogs/%23openstack-qa/%23openstack-qa.2022-05-24.log.html#t2022-05-24T15:31:51

Change-Id: Ie844887a798a006794729e733748ceaa5e06d2db