Use yoga stable constraint in tox to release 31.0.0
We are going to release Tempest new tag 31.0.0 to declare
the end of support for stable/ussuri. Pinning constraints
in tox.ini with stable/yoga constraint because Tempest is tested
with stable/yoga and master constraints. It did not use master constraints
as master dev cycle is in progress and so does their constraints also
is in progress and keep getting modified.

Further information:
 https://docs.openstack.org/tempest/latest/requirement_upper_constraint_for_tempest.html

Change-Id: Ief33c3ebb527b81ae8fbd222237ad5605fad045a