Wait for floating IP assocation in test_minimum_basic test
In bug 1923194 Tempest wasn't giving Nova a change to update its
network info cache after attaching a floating IP, leading to failures
when the floating IP association was asserted via the Nova API. The
fix was to add a waiter before doing the assertion. We were not using
that waiter in the scenario floating IP association code, leading to
bug 1980255. This patch starts doing that.

Related-bug: 1980255
Change-Id: I05e079237aec332925a88476e45071b34dc722c6