move tempest-integrated-compute-centos-9-stream to periodic
As agreed on the Nova weekly meeting [1], we won't put back C9S job to voting
and we'll rather put it on both experimental and periodic-weekly pipelines.

Change I2de4d90202fd7e735026150f79409d51f2d70e0e attempted to
move this job locally in nova however the integrated-gate-compute template
reintoduces the job to the check and gate pipelines.

This change updates the integrated-gate-compute templates to reflect the
nova teams wishes with regards to not running this job on every change
and only monitoring it via our weekly meeting.

[1] https://meetings.opendev.org/meetings/nova/2022/nova.2022-07-05-16.00.log.html#l-52

Change-Id: I39480ab239ac48d899bd23c00b7cf9a9485ba34b