Remove transfer-encoding headers from test
The reason to remove the headers from this test is because urllib3
already add this header on the response. Prior urllib3 1.26.6 the header
is duplicated (one header from tempest, and one header from urllib
itself). Which it is fine according to RFC2616, however, from swift
perspective, it will fail if multiple headers are sent [1].
This issue doesn't happen with urllib3 > 1.26.5 because it checks if the
transfer-encoding header is already present, and doesn't add it [2].
This test is failing on tripleo side that uses urllib3 1.26.5, which is
in accordance with tempest requirements and upper constraints from
openstack, so, removing the headers here fix the problem.

1 - https://github.com/openstack/swift/commit/5174b7f85dea42b6e0a50ee0a222637be0cea240
2 - https://github.com/urllib3/urllib3/commit/eb28aec79de922100b246f2ac1cd4625394ff70c

Change-Id: I418891d3a00e22effb394a553d8f40552da37713
Closes-Bug: #1982874