Use stable constraint in tox to release new tag for Zed
We are going to release Tempest new tag to declare the start of
support for stable/zed. So that new tag use stable/zed
constraint in the tox env.

For example: if anyone use Tempest new tag (32.0.0) in future say
1 year later then tox env also should use the zed contstraint for
compatibility instead master.

These need to be move back to master constraint once release is done.

Change-Id: I188a23bf77855358535986256b17fcc4cc6405f9