Fix image import tests for read-only stores
In order to test with one or more read-only store (such as an http
type store), we must exclude those from our list of stores we expect
to see when doing an import. This fixes both the waiter for the
"all stores" case, as well as our specific-store test from choosing
those stores for the list.

Change-Id: I8dd5e3256339ab1483e909fb5207d0da856e467e