Add replace location test
Replacing locations is a little strange, in that you "replace" the
whole list of locations, but only metadata for existing locations
is allowed, but outside the JSON PATCH checking. This adds a test
that probes the boundaries of that and makes sure we are allowed
(and disallowed) the right operations.

Change-Id: I753872656b990741ababfbdd64893746493b5e13