zuul: Increase timeout for c9s-fips job
tempest-centos9-stream-fips job has a long history of runs resulting to
TIMED_OUT [1]

Since the job defines some extra tasks for enabling FIPS, we hardly hit
the default timeout of 2 hours as shown in every failing build log.
If we really want this job, then we need to give it some extra space as
for the moment, apart from wasting CI resources into continuously failing
runs, we don't see any interest in it.

[1] https://zuul.openstack.org/builds?job_name=tempest-centos9-stream-fips

Change-Id: I5b04d69f74854725e0c11abe58e936fc6d18d183