Fix default values for variables in run-tempest role
With the new version of tox the parsing of postargs value has changed.
Previously an empty string was discarded by tox if it appeared in the
postargs value in contrast with the latest version of tox which
leaves the empty string intact.

The behavior of the latest tox version makes some jobs fail as some
variables in the run-tempest role have an empty string as a default
value. This leads in some cases to the execution of an incorrect tempest
command, for example:

tempest run --regex '[.*\bsmoke\b.*]' '' --concurrency=2

The patch changes the default value for blacklist_option,
exclude_list_option and tempest_test_regex variables to None to prevent
an empty string from being passed to postargs variable in tox.

Change-Id: I006e33c5ea0753b68cf1feb8580cb029209af9a8