Chunked GET request support
In one test, we are downloading the entire image (into memory) and
re-uploading it. That works when the image is 16MiB but not when it
is 1GiB. This adds support to the internal http client for chunked
downloads (similar to upload), makes the image client able to take
that flag, and finally makes the offending test do a chunked upload/
download streaming operation.

Note this un-skips the test, effectively reverting a6b7e334c
because the test should no longer consume large amounts of memory.

Related-Bug: #2002951
Change-Id: I31e537538a1862e71091aa470da3b8e9c799bf15