Fix escaping of regular expressions in tox.ini
With the new version of tox, the interpretation of escaping in the
command section changed [1]. For example, the `tox -e smoke` runs the
following command to execute tempest tests when tox==4.3.5. is installed:

tempest run --regex '[.*\bsmoke\b.*]'
(missing '\' in front of '[', ']') ^^

The issue can be fixed either by adding another backslash ('\\[') or
moving the regular expressions to separate variables. The first solution
won't, however, be backwards compatible with the older version of tox
therefore this patch uses the second approach.

[1] https://tox.wiki/en/latest/config.html#substitutions

Change-Id: Ie5ea7f9305c7e7611f205d2e0703849bd6321a68