Fix retry_bad_request() context manager
When I converted this from a decorator to a context manager, I didn't
remove the fn parameter, nor was it even doing the thing I expected
on the second and later iterations because it doesn't fail in my local
environment. Apparently we're not running this test in
the tempest gate, so this adds to experimental so we can at least
run it on command.

Change-Id: Ia72b50f7f7bf64fe0ddd3f1a415b1807ff264b66