Separate the extra tests to run in a separate job
Recently we are seeing a lot of job timeout(bug#2004780)
and we see many tests taking time and also number of tests
increasing over time. This is to prepare the list of extra tests
(here extra tests means the tests which are covered by the other
API or scenario tests) which we do not need to run in every
integrated jobs. Instead, we can run them in a separete job(s).

Currently I am adding admin (except keystone) and negative tests
in the 'extra tests' list but we can add more tests here which
we think are covered in some other tests.

As negative tests are important for interop, adding those extra
tests coverage for stable branch job also but running them in
periodic run only.

Related-Bug: #2004780
Change-Id: Id02221df0d6180519751c63e890851bd59fdafa0