Use stable constraint in tox to release new tag for 2023.1
We are going to release Tempest new tag to declare the start of
support for stable/2023.1. So that new tag use stable/2023.1
constraint in the tox env.

For example: if anyone use Tempest new tag in future say
1 year later then tox env also should use the 2023.1 contstraint for
compatibility instead master.

These need to be move back to master constraint once release is done.

Change-Id: Iedaca324ab239cdb102b28f0eb40a9d160d1591f