Avoid wait for vol detach in live-migrate tests
This extends the recent theme of making sure the tearDown process does
not hang and ultimately fail the test because it cannot detach a
volume from a guest. We see (and have been seeing for a long time)
the live migration tests specifically fail in this way in the gate.

Change-Id: Ib1065c146ab2f927c23fc8a0b73d7040516a19a6
Related-Bug: #199232