Fix server cleanup in test_multiple_create test
The above-mentioned test uses ServersClient.delete_server function to
delete servers. However, this function does not wait for the successful
deletion of the servers which sometimes leads to an error (see linked
bug).

This patch changes the test so that it uses
BaseV2ComputeTest.delete_server function which waits for the successful
deletion of the server.

Closes-Bug: #1954535
Change-Id: Ic1ed48fda03ce3c13198d2dd299b57d68b06e63d