Update "test_list_migrations_in_flavor_resize_situation" to indicate the failing request on timeout.
This is one of those tests that for some reason or the other fail randomly on our CI systems. Problem is, that the error message it fails with is not very helpful, as it just indicates that a timeout happened but not much else. This change seeks to fix that by including the request id that led to the timeout. This way, once the test fails on a build, it will be possible to search the log files for this request and track the underlying problem.

Closes-Bug: #2009701
Change-Id: Ic1ffd62f9d6ed0d69467f678c1820cddbb1fda12