Allow wait_until=SSHABLE in scenario manager
If we can, we should always wait for servers to be ssh'able before we
try to do things to them, especially attach volumes. This improves
stability of the tests, at the expense of performance.

This also makes some volume-having scenario tests wait for SSHABLE.
The ones that don't are already using ssh to do things like create
and check timestamps, so they need not be changed to still have this
behavior.

Change-Id: I169d8d119758e4f5e79c42cfd458577a67b13fc5