Allow passing validation_resources in scenario
In the scenario manager, we need to pass through validation parameters
if they are passed to us. If not, we need to do the needful for
SSHABLE targets, but some tests need to pass in their own. Note that
the scenario manager create_server() already has a validatable
parameter, but it does *not* pass that through to create_test_server(),
so we also need to fix that up for those cases to avoid tripping
over the assertion requirement that it be provided.

Change-Id: I44b3deba70e3f33f1287a6b3f28c21da5492cc04