Remove closure _unshelve_server not useful
The "_unshelve_server" method is not effective since it is executed
after tearDown, as it is scheduled with addCleanup. During the tearDown
process, the "server_check_teardown" method waits for a server to be
in the "ACTIVE" state. If a timeout occurs, the server is destroyed
to create a new one and continue the test safely.

Therefore, either the test exits tearDown with an "ACTIVE" server, in
which case "_unshelve_server" will do nothing since it checks for a
"SHELVED" server, or the server is destroyed upon exiting tearDown,
and "_unshelve_server" raises an exception, as we can see in the
stack trace for issue #2016957.

Related-Bug: #2016967
Change-Id: Id34c82b001c77ba65ae987dc5eb6b925bc1b279f