Fix test_delete_locations to be performed by admin
Delete image location operation is allowed to admin only
by default
- https://github.com/openstack/glance/blob/428295969836a2a0bad0ae0593ea1a890e155898/glance/policies/image.py#L150

But tempest test was performing it with non admin creds
- https://github.com/openstack/tempest/blob/4054e13e65f7d11c9097946a36976bf9c6edea45/tempest/api/image/v2/test_images.py#L935

While enabling the Glance RBAC new defaults in devstack (
everything run with RBAC new defaults), this test start failing
with 403 error

- https://review.opendev.org/c/openstack/devstack/+/883601
- https://storage.gra.cloud.ovh.net/v1/AUTH_dcaab5e32b234d56b626f72581e3644c/zuul_opendev_logs_2dd/883601/1/check/nova-ceph-multistore/2dd523c/testr_results.html

It was not failing before because this test was only enabled in
nova-ceph-multistore job which used old policy and old policy allow
everyone to perform this operation. tempest-full-enforce-scope-new-defaults
job which enable new defaults did not run this test to did not capture this
issue.

Change-Id: I6cbf95caf782bde1b11494b13e9d8d39ab9f3080