Increase concurrency for tempest-slow jobs
tempest-slow jobs are set with concurrency 2 which
means only two cpu are used to run the test. There
is no reason to restrict the tempest slow job to
two test runner worker. Maybe this is causing this
job to take long time and now a days it is timeout
a lot
- https://zuul.opendev.org/t/openstack/builds?job_name=tempest-slow-py3&skip=0

Let's unset the concurrency and stestr distribute the tests
to run by counting the number of cpu on test node.

Change-Id: I3d5f84279859a5f6ca5ce9ab81041974fd0f6c2f