Increase the default concurrency for tempest run
run-tempest role which run tempest test for all tempest
based job limit the concurrency (number of worker used to
run the tests) to number-of-cpu/2 (if number of cpu > 3)
which limit the jobs to use half of the available cpu to
run the tests. We are seeing a lot of timeout now a days
and increasing the test runner worker can possibily solve
the timeout issue.

This commit modifies the logic of default concurrency to
set as number-of-cpu -2 so that we will be utlizing the
available cpu in more efficient way.

Related-Bug: #2004780

Change-Id: I8beeedade3098ad346d3a307b94bc3b254cde90a