Split up ServerActionsTestJSON for better mixing
Since we group tests in a single class on a given testr worker, a very
large/slow class will cause a bunch of wall time to pile up on a
single worker, making our distribution less efficient. This splits up
the tests in ServerActionsTestJSON into two other classes (and
refactors the helpers to a common base class). The interop tests that
can't move stay in a class of the same name, while the rest move into
two "other" classes for better distribution.

On my system, this results in an almost perfect distribution of wall
time into three buckets and goes from about 15 minutes serial to just
over five minutes in parallel.

Change-Id: Ib0318bb3c55d05891f209af259e706fa66a3081f