Optimize the resource creation and status waiting in scenario tests
In many scenario tests, resource either same or different services
are created serially which is ok at some extend as most of the
resource creation is async operation. But when we wait for resource
status serially for example volume to reach to 'avaiable' state, server
to become 'SSHABLE' then it add a lot of time in test execution.

This commit try to make resources creation and their status waiting
in parallel operation so that while we wait for volume to be reach in
'available' state we create server and same time we wait for server
to be ACTIVE or SSHABLE.

Related-Bug: #2004780
Change-Id: I30f2a44d94dc1e94aefeefecd4dc4e25bdda7a72