Correct _get_console() to get console of correct server
_get_console() method always try to get the console
for the self.server which is created in setup but
test test_get_console_output_server_id_in_shutoff_status
create its own server and use _get_console method.

This way test create its own server but try to get
console of different server

Due to above issue test_get_console_output_server_id_in_shutoff_status
was always wrong which used to get the console for the server
created in the setup method and is in active state. This test
never tried to get the console of the shutoff server. After
correcting this tests, it fail 100% because Nova return 404
for shutoff server console. It is still under discussion that
what should be the expected behaviour in this case. Meanwhile
we figure that out, skipping this test to unblock the gate.

Related-Bug: #2028851
Change-Id: I86759f6b998f916b22d9ae0321b656598266be24