Skip scenario tests early to avoid unnecessary setup
This is change for the volume scenario test to
skip them early.

When we skip the test class using skip_checks(), it check
the conditions and skip the test class at first step
without creating any keystone credentials. But when
tests are skipped with other decorator at test level then
it does create keystone credentials, setup network resources
and service clients.

This will mostly help neutron gate where these volume
tests will be skipped in the initial stage only and will
not create the keystone and network resources.

One good example is TestEncryptedCinderVolumes which is skipped
 - https://zuul.openstack.org/build/babcc06f24764a408ed77702365b4c5b/log/job-output.txt#28695

But it still does the resources setup
- https://zuul.openstack.org/build/babcc06f24764a408ed77702365b4c5b/log/controller/logs/tempest_log.txt#6374-6450

Related-Bug: #2004780
Change-Id: I59cd39c20b995bf2ed2f58f4522743c3ca51b516