Fix cleanup for volume backup tests
Volume backup tests create a container when swift is used as
a backup driver. This causes a failure of an object storage
test [1] as it expects no containers being present before
the testing.

This patch fixes the cleanup of volume backup tests by making
sure that containers created during testing are deleted.

The containers cleanup is called only when swift is present.
When Swift is not used as a back up driver but is present,
the cleanup tries to delete the containers anyway. This will
not cause a failure as the delete_containers function catches
the NotFound [2] error.

[1] https://opendev.org/openstack/tempest/src/tempest/api/object_storage/test_account_services.py#L67
[2] https://opendev.org/openstack/tempest/src/tempest/api/object_storage/base.py#L50

Closes-Bug: #2028671
Change-Id: If94facd5a926f7eadd092dfc8f0368d8e4b8d630