Adding tempest prefix option in tempest.conf
Adding a new config option in tempest.conf as the prefix for the random
name resources with this new option all the resources created by
tempest will have the prefix defined by resource_name_prefix option
in tempest.conf. If not defined the default is tempest.

Tempest cleanup CLI will then use this config option to cleanup only the
resources that match the prefix. Make sure this prefix does not match with
the resource name you do not want Tempest cleanup CLI to delete.

This is a patch created after the conversation on
https://review.opendev.org/c/openstack/tempest/+/879923?tab=comments

Related-Bug: #1945082
Change-Id: I663a6307a5aa63299d20a304188efe6d29aea9d3