Add '--slowest' option to 'tempest run'
A common pattern in some projects running tests with stestr is to
output the slowest tests from the test run after it finishes by running
'stestr slowest' after 'stestr run'.

We can do something similar in 'tempest run' with a '--slowest' option
that will run 'stestr slowest' when passed.

This addition is inspired by seeing TIMED_OUT job failures in the gate
with tests run using 'tempest run'. We could consider adding the
'--slowest' option to those command lines to make it easier to identify
the slowest tests in the job logs and potentially work on improving
their runtimes.

Change-Id: Ia88dabcb41d56d4246337ba67c140a93342ad6ab