cli auth: only inject project name if provided
One can create and need to execute execute functional
tests which may operate with a different scope, i.e. without
a project name. This was not previously functional as the
tempest method would always inject a project name when
none was provided, which would then result in commands
being executed with "--os-project-name None" which would
result in the command failing.

Change-Id: Ieeed191addb277f0c39c7a11f1f77034915be7fc