Add wait for location import task
When uploading a volume to image in optimized path, we are working
to get support of new location APIs.
The new location APIs are async and we need to wait for the task
to finish before performing the cleanup else we will fail.
This patch adds a waiter for location import task during upload volume.

Change-Id: Ia19e0bcc318ba7c4e71c2f36343b783431743a3e