Use stable constraint in tox to release new tag for 2024.1
We are going to release Tempest new tag to declare the start of
support for stable/2024.1. So that new tag use stable/2024.1
constraint in the tox env.

For example: if anyone use Tempest new tag in future say
1 year later then tox env also should use the 2024.1 contstraint for
compatibility instead master.

Also, making ironic-tempest-bios-ipmi-direct-tinyipa as
non voting to merge his change.

These need to be reverted once release is done.

Change-Id: I5ee1ce6b3ffbbb1f47e448fc15d540f74b21684f