Enable tempest cleanup with record_resources in jobs
This enables tempest cleanup with record_resources in the tempest-all job
to create a list of all the resources created during the tempest run.
Then the list will be used to run 'tempest cleanup' to clean only
the resources in the list.

Also, enable the dry cleanup with record_resources in tempest-extra-tests
job.

Change-Id: Ia64a1c09d4b5e9815ff5926161d550b56433ec64