Bump http_timeout in devstack-tempest
In the gate we occasionally see failures like the following:

  urllib3.exceptions.ReadTimeoutError:
    HTTPSConnectionPool(host='10.209.0.62', port=443):
    Read timed out. (read timeout=60)

  urllib3.exceptions.ReadTimeoutError:
    HTTPSConnectionPool(host='10.209.99.246', port=9696):
    Read timed out. (read timeout=60)

where Tempest gives up waiting for an HTTP response after 60 seconds.

It doesn't appear to be related only to one service, as it has been
seen when calling Nova API or Neutron API (the port=9696 example is a
call to Neutron API).

We could try bumping the http_timeout from the default 60 seconds to 90
seconds to see if it helps in slow environments.

Related-Bug: #1999893

Change-Id: I2b2a6a4dee40ee86ca89f6f7f9a83409f8312617